moving files from src/server to src and fixing the paths
moving:
achievements folder
auth folder
prisma folder
