+++ src/achievements/achievements.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { AchievementsService } from './achievements.service';
-import { AchievementsController } from './achievements.controller';
 import { PrismaService } from '../prisma/prisma.service';
+import { AchievementsController } from './achievements.controller';
+import { AchievementsService } from './achievements.service';
 
 @Module({
   controllers: [AchievementsController],
