+++ src/auth/jwt/refresh.guard.ts
@@ -1,5 +1,5 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
-import { PrismaService } from 'src/server/prisma/prisma.service';
+import { PrismaService } from 'src/prisma/prisma.service';
 
 
 @Injectable()
